Fix Algorithm train and implement calcPop.
Train was implemented in a way that didn't use calcPop at all. We had to reformat the conditionals to properly reflect the scala version.

CalcPop was somehow neglected, so we did a line-by-line translation of the scala code, using the same calcAllFunction class that was a helper in calcAll (same functionality required).

James Mackey